+++ src/main/java/com/top/effitopia/service/CheckoutService.java
@@ -5,16 +5,9 @@ import com.top.effitopia.dto.*;
 import java.util.List;
 
 public interface CheckoutService {
-//    PageResponseDTO<CheckoutDTO> getList(PageRequestDTO pageRequestDTO);
-//    boolean save(CheckoutDTO checkoutDTO, CheckoutQuestionDTO checkoutQuestionDTO, CheckoutAnswerDTO checkoutAnswerDTO);
-//    boolean save(CheckoutDTO checkoutDTO);
-//
-//    boolean save(CheckoutDTO checkoutDTO, List<CheckoutAnswerDTO> answers);
-//
-//    List<CheckoutQuestionDTO> getAllQuestions();
     PageResponseDTO<CheckoutDTO> getList(PageRequestDTO pageRequestDTO);
 
     boolean save(CheckoutDTO checkoutDTO, List<CheckoutAnswerDTO> checkoutAnswers);
 
-    CheckoutDTO getOne(Long checkoutId);
+    CheckoutDTO getOne(Integer checkoutId);
 }
